// src/typing/models/LoginCredential.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { UserInfo } from './UserInfo';
import {
    UserInfoFromJSON,
    UserInfoFromJSONTyped,
    UserInfoToJSON,
    UserInfoToJSONTyped,
} from './UserInfo';

/**
 * 
 * @export
 * @interface LoginCredential
 */
export interface LoginCredential {
    /**
     * 
     * @type {string}
     * @memberof LoginCredential
     */
    username?: string;
    /**
     * 
     * @type {string}
     * @memberof LoginCredential
     */
    password?: string;
    /**
     * 
     * @type {string}
     * @memberof LoginCredential
     */
    email?: string;
    /**
     * 
     * @type {boolean}
     * @memberof LoginCredential
     */
    setPassword?: boolean;
    /**
     * 
     * @type {UserInfo}
     * @memberof LoginCredential
     */
    userInfo?: UserInfo;
}

/**
 * Check if a given object implements the LoginCredential interface.
 */
export function instanceOfLoginCredential(value: object): value is LoginCredential {
    return true;
}

export function LoginCredentialFromJSON(json: any): LoginCredential {
    return LoginCredentialFromJSONTyped(json, false);
}

export function LoginCredentialFromJSONTyped(json: any, ignoreDiscriminator: boolean): LoginCredential {
    if (json == null) {
        return json;
    }
    return {
        
        'username': json['username'] == null ? undefined : json['username'],
        'password': json['password'] == null ? undefined : json['password'],
        'email': json['email'] == null ? undefined : json['email'],
        'setPassword': json['setPassword'] == null ? undefined : json['setPassword'],
        'userInfo': json['userInfo'] == null ? undefined : UserInfoFromJSON(json['userInfo']),
    };
}

export function LoginCredentialToJSON(json: any): LoginCredential {
    return LoginCredentialToJSONTyped(json, false);
}

export function LoginCredentialToJSONTyped(value?: LoginCredential | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'username': value['username'],
        'password': value['password'],
        'email': value['email'],
        'setPassword': value['setPassword'],
        'userInfo': UserInfoToJSON(value['userInfo']),
    };
}